% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_mcmc_diagnostics.R
\name{divergent_transitions}
\alias{divergent_transitions}
\title{Return number of divergent transitions}
\usage{
divergent_transitions(fit, absolute = FALSE, ...)
}
\arguments{
\item{fit}{An brmsfit or stanfit object.}

\item{absolute}{True if the number of divergent transitions should be
returned. Otherwise the fraction of divergent transitons to
total post-warmup samples is returned as default.}

\item{...}{Potential future arguments.}
}
\value{
Number of divergent transitions
}
\description{
Return number of divergent transitions
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
divergent_transitions(fit)
}
