% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_metrics.R
\name{elpd_loo}
\alias{elpd_loo}
\title{Loo expected log pointwise predictive density and effective number of
parameters}
\usage{
elpd_loo(fit, ...)
}
\arguments{
\item{fit}{A brmsfit object}

\item{...}{Additional arguments for \code{\link[loo]{loo}}}
}
\value{
A named list containing elpd_loo and p_loo estimates, standard errors
        and pointwise estimates
}
\description{
Mainly a convenient wrapper for \code{\link[loo]{loo}} to get the quantities
bayesim is interested in.
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
elpd_loo(fit)
}
