% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_metrics.R
\name{elpd_newdata}
\alias{elpd_newdata}
\title{Expected log pointwise predictive density based on new data}
\usage{
elpd_newdata(fit, newdata)
}
\arguments{
\item{fit}{a brmsfit object}

\item{newdata}{A data.frame containing the new dataset ob which to evaluate
elpd on}
}
\value{
A named list containing the elpd estimate, standard error and
        pointwise estimate.
}
\description{
Expected log pointwise predictive density based on new data
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
elpd_newdata(fit, newdata = list(y = rnorm(1000)))
}
