% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_mcmc_diagnostics.R
\name{ess_bulk.brmsfit}
\alias{ess_bulk.brmsfit}
\title{Bulk effective sample size (bulk-ESS)}
\usage{
\method{ess_bulk}{brmsfit}(fit, variables)
}
\arguments{
\item{fit}{brmsfit object.}

\item{variables}{Vector of variable names of interest.}
}
\value{
Named list of bulk effective sample size for each variable
}
\description{
Bulk effective sample size (bulk-ESS)
}
\examples{
fit <- brms::brm(y ~ 1, data = rnorm(1000))
ess_bulk(fit, c("b_Intercept", "sigma"))
}
