% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_mcmc_diagnostics.R
\name{ess_tail.brmsfit}
\alias{ess_tail.brmsfit}
\title{Tail effective sample size (tail-ESS)}
\usage{
\method{ess_tail}{brmsfit}(draws, variables)
}
\arguments{
\item{variables}{Vector of variable names of interest.}

\item{fit}{brmsfit object.}
}
\value{
Named list of tail effective sample size for each variable
}
\description{
Tail effective sample size (tail-ESS)
}
\examples{
fit <- brms::brm(y ~ 1, data = rnorm(1000))
ess_tail(fit, c("b_Intercept", "sigma"))
}
