% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_mcmc_diagnostics.R
\name{ess_tail.draws}
\alias{ess_tail.draws}
\title{Tail effective sample size (tail-ESS)}
\usage{
\method{ess_tail}{draws}(draws, variables)
}
\arguments{
\item{draws}{posterior draws object, eg. from
\code{\link[posterior]{as_draws}}}

\item{variables}{Vector of variable names of interest.}
}
\value{
Named list of tail effective sample size for each variable
}
\description{
Tail effective sample size (tail-ESS)
}
\examples{
fit <- brms::brm(y ~ 1, data = rnorm(1000))
ess_tail(posterior::as_draws(fit), c("b_Intercept", "sigma"))
}
