% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_mcmc_diagnostics.R
\name{pareto_k_values}
\alias{pareto_k_values}
\title{Pareto-khat estimate values}
\usage{
pareto_k_values(x)
}
\arguments{
\item{x}{brmsfit or psis object}
}
\value{
List of all pareto-khat estimate values
}
\description{
Pareto-khat estimate values
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
bad_pareto_ks(fit)
}
