% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_posterior_probability.R
\name{pos_prob}
\alias{pos_prob}
\title{Title}
\usage{
pos_prob(posterior_draws)
}
\arguments{
\item{posterior_draws}{posterior::extract_variable_matrix for relevant parameter}
}
\description{
Title
}
