% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_mcmc_diagnostics.R
\name{post_warmup_samples}
\alias{post_warmup_samples}
\title{Post warmup samples}
\usage{
post_warmup_samples(fit, ...)
}
\arguments{
\item{fit}{An brmsfit or stanfit object.}

\item{...}{Potential future parameters.}
}
\value{
The total number of post-warmup samples.
}
\description{
Post warmup samples
}
\examples{
post_warmup_samples(brms::brm(y ~ 1, data = list(y = rnorm(100)), iter = 500))
}
