% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_bias.R
\name{posterior_bias}
\alias{posterior_bias}
\title{Title}
\usage{
posterior_bias(posterior_draws, true_value, ...)
}
\arguments{
\item{posterior_draws}{posterior::extract_variable_matrix for relevant parameter}

\item{true_value}{The true parameter value to calculate the bias against.}
}
\description{
Title
}
