% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_mean.R
\name{posterior_mean}
\alias{posterior_mean}
\title{Title}
\usage{
posterior_mean(posterior_draws)
}
\arguments{
\item{posterior_draws}{posterior::extract_variable_matrix for relevant parameter}
}
\description{
Title
}
