% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_quantiles.R
\name{posterior_quantiles}
\alias{posterior_quantiles}
\title{Calculate quantiles based on posterior samples for a given parameter.}
\usage{
posterior_quantiles(x, variable, probs, ...)
}
\arguments{
\item{x}{brmsfit objec or posterior draws object, eg. from
\code{\link[posterior]{as_draws}}.}

\item{variable}{Name of the parameter of interest.}

\item{probs}{Vector of quantiles, passed to \code{\link{quantiles}} function.}

\item{...}{Additional arguments passed to \code{\link{quantiles}}.}
}
\value{
Named list of quantiles.
}
\description{
Calculate quantiles based on posterior samples for a given parameter.
}
\examples{
fit <- brms::brm(y ~ 1, data = rnorm(1000))
posterior_quantiles(fit, "b_Intercept", c(0.025, 0.5, 0.975))

}
