% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_sample_mae.R
\name{posterior_sample_mae}
\alias{posterior_sample_mae}
\title{Title}
\usage{
posterior_sample_mae(posterior_draws, true_value, ...)
}
\arguments{
\item{posterior_draws}{posterior::extract_variable_matrix for relevant parameter}

\item{true_value}{The true parameter value to calculate the mae against.}

\item{...}{}
}
\description{
Title
}
