% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_sd.R
\name{posterior_sd}
\alias{posterior_sd}
\title{Title}
\usage{
posterior_sd(posterior_draws)
}
\arguments{
\item{posterior_draws}{posterior::extract_variable_matrix for relevant parameter}
}
\description{
Title
}
