% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summary.R
\name{posterior_summary}
\alias{posterior_summary}
\title{Calculate a summary statistic of the posterio for a given parameter.}
\usage{
posterior_summary(x, variable, fun)
}
\arguments{
\item{x}{brmsfit object, posterior draws list from
\code{\link[posterior]{as_draws}} or matrix from
\code{\link[posterior]{extract_variable_matrix}}.}

\item{variable}{Name of the parameter of interest.}

\item{fun}{Summary function, eg mean, median or sd.}
}
\value{
The calculated summary of the posterior.
}
\description{
Calculate a summary statistic of the posterio for a given parameter.
}
\examples{
fit <- brms::brm(y ~ 1, data = rnorm(1000))
posterior_summary(fit, "b_Intercept", mean)

}
