% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_metrics.R
\name{r2}
\alias{r2}
\title{Calculate R^2 for given y and yrep.}
\usage{
r2(y, yrep, weights = NULL)
}
\arguments{
\item{y}{Vector of observed values}

\item{yrep}{Vector of predicted Values}

\item{weights}{PSIS weights}
}
\value{
R^2 for the given y and yrep vectors
}
\description{
If psis-weights are supplied, the corresponding psis-R^2 is returned.
}
\examples{
r2(c(1:10), c(5:14))
}
