% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_metrics.R
\name{r2_newdata}
\alias{r2_newdata}
\title{R² for new data}
\usage{
r2_newdata(fit, newdata)
}
\arguments{
\item{fit}{A brmsfit object.}

\item{newdata}{A data.frame containing the new dataset ob which to evaluate
elpd on}
}
\value{
Named list with R² estimate, standard error and pointwise estimate.
}
\description{
R² for new data
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
r2_newdata(fit, newdata = list(y = rnorm(1000)))
}
