% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_mcmc_diagnostics.R
\name{rhat.draws}
\alias{rhat.draws}
\title{Rhat convergence diagnostic}
\usage{
\method{rhat}{draws}(draws, variables)
}
\arguments{
\item{draws}{posterior draws object, eg. from
\code{\link[posterior]{as_draws}}}

\item{variables}{Vector of variable names of interest.}
}
\value{
Named list of rhat values for each variable.
}
\description{
Rhat convergence diagnostic
}
\examples{
fit <- brms::brm(y ~ 1, data = rnorm(1000))
rhat(posterior::as_draws(fit), c("b_Intercept", "sigma"))

}
