% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_metrics.R
\name{rmse_loo}
\alias{rmse_loo}
\title{PSIS-LOO rmse}
\usage{
rmse_loo(fit, psis_object = NULL)
}
\arguments{
\item{fit}{A brmsfit object}

\item{psis_object}{An optional psis object to prevent recalculation.}
}
\value{
Named list with rmse estimate, standard error and pointwise estimate.
}
\description{
PSIS-LOO rmse
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
rmse_loo(fit)
}
