% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_metrics.R
\name{rmse_newdata}
\alias{rmse_newdata}
\title{rmse based on new data}
\usage{
rmse_newdata(fit, newdata)
}
\arguments{
\item{fit}{A brmsfit object}

\item{newdata}{A data.frame containing the new dataset ob which to evaluate
elpd on}
}
\value{
Named list with rmse estimate, standard error and pointwise estimate.
}
\description{
rmse based on new data
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
rmse_newdata(fit, newdata = list(y = rnorm(1000)))
}
