% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstar.R
\name{rstar_w}
\alias{rstar_w}
\title{Calculate R* convergence diagnostic}
\usage{
rstar_w(x, ...)
}
\arguments{
\item{x}{brmsfit object or posterior draws list from
\code{\link[posterior]{as_draws}}}

\item{...}{Additional arguments for \code{\link[posterior]{rstar}}}
}
\value{
A numeric vector of length 1 (by default) or length nsimulations
 (if uncertainty = TRUE).
}
\description{
A convenient wrapper around \code{\link[posterior]{rstar}}
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
rstar_w(fit)
}
