% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_error.R
\name{sample_error}
\alias{sample_error}
\title{Calculate an error based on posterior samples for a given parameter.}
\usage{
sample_error(x, variable, error, reference)
}
\arguments{
\item{x}{brmsfit object, posterior draws list from
\code{\link[posterior]{as_draws}} or matrix from
\code{\link[posterior]{extract_variable_matrix}}.}

\item{variable}{Name of the parameter of interest.}

\item{error}{Identifier of the error measure to calculate.}

\item{reference}{Reference value to calculate the error against.}
}
\value{
The calculated error of the posterior.
}
\description{
Calculate an error based on posterior samples for a given parameter.
}
\examples{
vfit <- brms::brm(y ~ 1, data = rnorm(1000))
sample_error(fit, "b_Intercept", "bias", 0)

}
