% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_mcmc_diagnostics.R
\name{sampling_time}
\alias{sampling_time}
\title{MCMC sampling times}
\usage{
sampling_time(fit, absolute = FALSE, ...)
}
\arguments{
\item{fit}{An brmsfit or stanfit object.}

\item{absolute}{TRUE if the absolute times should be returned, FALSE
(default) if time per sample should be returned.}

\item{...}{Potential future parameters.}
}
\value{
List containing warmup and sampling time
}
\description{
MCMC sampling times
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
#' fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
bad_pareto_ks(fit)
}
