% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_posterior_quantile.R
\name{true_posterior_quantile}
\alias{true_posterior_quantile}
\title{Title}
\usage{
true_posterior_quantile(posterior_draws, true_value, ...)
}
\arguments{
\item{posterior_draws}{posterior::extract_variable_matrix for relevant parameter}

\item{...}{}
}
\description{
Title
}
