% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_distance.R
\name{variable_distance}
\alias{variable_distance}
\title{Calculate a distance metric of the posterior for a given variable and
reference value.}
\usage{
variable_distance(x, variables, references, fun)
}
\arguments{
\item{x}{brmsfit object or posterior draws object, eg. from
\code{\link[posterior]{as_draws}}.}

\item{variables}{Vector of variable names of interest.}

\item{references}{Vector of reference values for the variables of interest.}

\item{fun}{Distance function, eg \code{\link{variable_rmse}}.}
}
\value{
The calculated summary of the posterior.
}
\description{
Calculate a distance metric of the posterior for a given variable and
reference value.
}
\examples{
fit <- brms::brm(y ~ 1, data = rnorm(1000))
variable_distance(
  fit,
  c("b_Intercept", "sigma"),
  c(0, 1),
  fun = variable_bias
)
}
