% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_distance.R
\name{variable_mae}
\alias{variable_mae}
\title{Posterior sample mae for a single variable.}
\usage{
variable_mae(draws, reference)
}
\arguments{
\item{draws}{Posterior draws for the variable of interest.}

\item{reference}{Reference value to calculate the bias against.}
}
\value{
Posterior sample mae for a single variable.
}
\description{
Posterior sample mae for a single variable.
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
variable_mae(posterior::extract_variable(fit, "b_Intercept"), 0)
}
