% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_summary.R
\name{variable_pos_prob}
\alias{variable_pos_prob}
\title{Positive probability mass of the posterior samples for a single variable.}
\usage{
variable_pos_prob(draws)
}
\arguments{
\item{draws}{Posterior draws for the variable of interest.}
}
\value{
Positive probability mass of the posterior samples for a single
        variable.
}
\description{
Positive probability mass of the posterior samples for a single variable.
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
pos_prob(posterior::extract_variable(fit, "b_Intercept"))
}
