% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_summary.R
\name{variable_summary}
\alias{variable_summary}
\title{Calculate a summary statistic of the posterior for a given variable.}
\usage{
variable_summary(x, variables, fun, ...)
}
\arguments{
\item{x}{brmsfit object or posterior draws object, eg. from
\code{\link[posterior]{as_draws}}}

\item{variables}{Vector of variable names of interest.}

\item{fun}{Summary function, eg \code{\link{mean}}.}

\item{...}{Potential future parameters.}
}
\value{
The calculated summary of the posterior.
}
\description{
Calculate a summary statistic of the posterior for a given variable.
}
\examples{
fit <- brms::brm(y ~ 1, data = rnorm(1000))
variable_summary(fit, c("b_Intercept", "sigma"), fun = mean)
}
