% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_distance.R
\name{variable_true_percentile}
\alias{variable_true_percentile}
\title{Percentile of the reference value in the draws.}
\usage{
variable_true_percentile(draws, reference)
}
\arguments{
\item{draws}{Posterior draws for the variable of interest.}

\item{reference}{Reference value to calculate the bias against.}
}
\value{
Percentile of the reference value in the draws.
}
\description{
Percentile of the reference value in the draws.
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
variable_mse(posterior::extract_variable(fit, "b_Intercept"), 0)
}
