% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_mcmc_diagnostics.R
\name{divergent_transitions}
\alias{divergent_transitions}
\title{Return number of divergent transitions}
\usage{
divergent_transitions(fit)
}
\arguments{
\item{fit}{brmsfit object}
}
\value{
Number of divergent transitions
}
\description{
Return number of divergent transitions
}
\examples{
fit <- brms::brm(y ~ 1, data = list(y = rnorm(1000)))
divergent_transitions(fit)
}
