% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_metrics.R
\name{get_y.default}
\alias{get_y.default}
\title{Rhat convergence diagnostic}
\usage{
\method{get_y}{default}(x, ...)
}
\arguments{
\item{x}{Any non-supported fit object.}

\item{...}{}
}
\value{
Warning that the fit model type is not supported.
}
\description{
Rhat convergence diagnostic
}
