% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_lookup.R
\name{metric_lookup.brmsfit}
\alias{metric_lookup.brmsfit}
\title{Access metrics with string identifiers}
\usage{
\method{metric_lookup}{brmsfit}(
  identifier,
  fit,
  posterior_draws = NULL,
  posterior_variable_matrix = NULL,
  testing_data = NULL,
  variables = NULL,
  references = NULL,
  threshold = 0.7,
  ...
)
}
\arguments{
\item{identifier}{A string that identifies a supported function.}
}
\value{
The function corresponding to the identifier string.
}
\description{
This function is mostly a helper function that maps identifier strings to
metrics for convenient use in some contexts.
}
