% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_quantiles.R
\name{posterior_quantiles}
\alias{posterior_quantiles}
\title{Title}
\usage{
posterior_quantiles(posterior_draws, prob)
}
\arguments{
\item{posterior_draws}{posterior::extract_variable_matrix for relevant parameter}

\item{prob}{Quantile}
}
\description{
Title
}
